Display the most recent telegram alert instead of a hard-coded post number
