+++ apiClient/AlertsApi.js
@@ -1,0 +1,7 @@
+import { apiGetFetch } from "./apiFetch";
+
+export const fetchMissileAlerts = (lang, count) =>
+  apiGetFetch(`/api/v1/missileAlerts`, {
+    lang,
+    count,
+  });
