+++ pages/alerts.js
@@ -1,12 +1,10 @@
-import { useState, useRef, useEffect } from "react";
-import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import Hero from "../components/Hero";
 import Layout from "../components/Layout";
 import MissileAlerts from "../components/MissileAlerts";
 import TelegramAlertLinks from "../components/TelegramAlertLinks";
-import { localeToTelegramAlertChannel } from "../utils";
+import TelegramEmbed from "../components/TelegramEmbed";
 import features from "../configs/features";
 
 export const getStaticProps = async ({ locale }) => ({
@@ -17,22 +15,6 @@ export const getStaticProps = async ({ locale }) => ({
 
 const AlertsPage = () => {
   const { t } = useTranslation();
-  const router = useRouter();
-
-  const [scriptLoaded, setScriptLoaded] = useState(false);
-  const ref = useRef();
-
-  useEffect(() => {
-    if (scriptLoaded || !ref.current)
-      return;
-    const channel = localeToTelegramAlertChannel(router.locale);
-    const tag = document.createElement("script");
-    tag.src = "https://telegram.org/js/telegram-widget.js?15"
-    tag.setAttribute("data-telegram-post", `${channel}/47`);
-    tag.setAttribute("data-width", "100%");
-    ref.current.appendChild(tag);
-    setScriptLoaded(true);
-  });
 
   return (
     <Layout
@@ -54,7 +36,7 @@ const AlertsPage = () => {
             title={t("Choose your language")}
           />
         </div>
-        <div ref={ref} style={{ maxWidth: "400px", margin: "0 auto" }} />
+        <TelegramEmbed />
         {features.liveMissileAlerts &&
           <>
             <div>
